refactor Toolbar: move subcomponents up
