+++ client/lib/ContextForm/template.ejs
@@ -1,4 +1,3 @@
-
 <% if (filters.length > 0) { %>
   <h5>Filters:</h5>
   <ul class="filter-list">
