+++ client/lib/ContextForm/toFilterLabel.js
