+++ client/lib/ContextForm/toFilterTypeLabel.js
